requirements.txt
# Workspace version: 1.44.5
# Core dependencies
fastapi>=0.95.0,<0.96.0
python-multipart>=0.0.6,<0.1.0
uvicorn[standard]>=0.22.0,<0.23.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
starlette>=0.26.0,<0.27.0

# Authentication
pyjwt[crypto]>=2.6.0,<3.0.0

# Database
firebase-admin>=6.1.0,<7.0.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.10.0,<2.0.0

# Caching
redis>=4.5.0,<5.0.0

# Monitoring and observability
prometheus-client>=0.16.0,<0.17.0
psutil>=5.9.0,<6.0.0

# AI/ML
openai>=0.27.0,<0.28.0

# Utilities
requests>=2.28.0,<3.0.0
beautifulsoup4>=4.11.0,<5.0.0
python-jose>=3.3.0,<4.0.0
email-validator>=2.0.0,<3.0.0

# Development dependencies (not needed in production)
# databutton  # Commented out due to dependency conflicts
ipykernel
pytest>=7.3.1,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-asyncio>=0.21.0,<0.22.0
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.3.0,<2.0.0
flake8>=6.0.0,<7.0.0